requirements.txt
fastapi==0.58.0
uvicorn==0.11.5
tortoise-orm==0.16.13